Some refactoring and a little bit of cleanup here.
