+++ ReadUs/RedisConnectionConfiguration.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Globalization;
 using System.Web;
 using ReadUs.Exceptions;
 
@@ -26,22 +27,20 @@ public readonly struct RedisConnectionConfiguration
 
     public static implicit operator RedisConnectionConfiguration(Uri connectionString)
     {
-        if (string.Compare(RedisScheme, connectionString.Scheme, false) != 0)
-            throw new RedisConnectionConfigurationException(
-                $"The provided scheme `{connectionString.Scheme}` is invalid, it must be `{RedisScheme}`.");
-
+        if (string.Compare(RedisScheme, connectionString.Scheme, false, CultureInfo.InvariantCulture) != 0)
+        {
+            throw new RedisConnectionConfigurationException($"The provided scheme `{connectionString.Scheme}` is invalid, it must be `{RedisScheme}`.");
+        }
         var host = connectionString.DnsSafeHost;
         var port = connectionString.Port == 0 ? DefaultRedisPort : connectionString.Port;
 
         var queryEntries = HttpUtility.ParseQueryString(connectionString.Query);
 
-        var connectionsPerNode = 1;
-
         if (int.TryParse(queryEntries.Get(ConnectionsPerNodeKey) ?? "1", out var parsedConnectionsPerNode))
-            connectionsPerNode = parsedConnectionsPerNode;
-        else
-            throw new RedisConnectionConfigurationException("`connectionsPerNode` must be a valid integer.");
+        {
+            return new RedisConnectionConfiguration(host, port, parsedConnectionsPerNode);
+        }
 
-        return new RedisConnectionConfiguration(host, port, connectionsPerNode);
+        throw new RedisConnectionConfigurationException("`connectionsPerNode` must be a valid integer.");
     }
 }
